bin: Allow 'any' to disable version check
To deploy Compliant Kubernetes, one would generally include this repo as
a submodule, in which case tags might be unavailable. This commit allows
for the config to have the version 'any', essentially meaning "I, the
operator, take responsibility to ensure the version of the config and
the version of ck8s match".