apps sc: Remove broken ops grafana oidc config
The ops grafana url is not registered as a valid redirect url, and
therefore login with oidc will not work.

There are decisions to be made regarding if/how to limit access to the ops
grafana instance.
Therefore there is no clear fix to the broken configuration at the moment.

This commit removes the broken configuration.