bin: Use split config, default and override
This includes changes to bin to add config loading as a step for most of
the commands, generating a temporary merged config that is used inplace
of the previous one.

It also includes changes to how secrets and passwords are generated.
They are now generated as a part of init scripts, and can be forced to be
regenerated.

The init script will also perform a backup of configs and secrets before
updating them.

Sets up the pipeline to use these changes, and to generate artifacts for
the pipeline config.

Fixes for `grafana-ops` using an invalid property name, and
`kube-prometheus-stack-sc` which had an invalid config for alertmanager.