apps wc: add default securitycontext mutations for user namespaces
Adds the default mutation for user namespaces that are locked down with
the restricted pod security policy. The helm release is installed in the
default namespace. This means that the gatekeeper constraints are
applied to that namespace, which should be fine as no user should have
access to the default namespace.