bin: Validate config against a Schema
JSON Schema is a format for describing the structure of a JSON value,
meant to support validation of data.

This makes it usable for validating the structure of the cluster config
files since JSON and YAML have overlapping data models.

Being an open standard, there are multiple tools and libraries available
for doing validation, as well as generating documentation or boilerplate
code for type-aware languages.

The initial schema only covers the top level 'global' entry for easier
review, additional entries will be added in subsequent PRs

Schema validation for secrets.yaml currently acts on the sops-encrypted
file and thus only sees the basic structure.

Includes a tool to generate basic json (but yaml) schemas, although it
requires some post-processing so it is mostly useful to get a starting
point.

To validate, run `./bin/ck8s validate sc|wc`. If something is wrong it
should print the path and value that violates the schema to make it easy
to fix.

Would have been nice if yajsv could output the the values that fail
validation itself.

suggestions from code review

Co-authored-by: André Arnqvist <58822152+aarnq@users.noreply.github.com>