migration: Track version in cluster trough upgrades
This is meant to atomically record the state transition into a migration (by creation) and then out of it (by deletion).

The general apps version is recorded at the end, and compared to the config version on next upgrade.

It should ensure that the config that is used to prepare is also used to apply, by way of storing a timestamp in the
config and the cluster and later ensuring that they match

Should ensure the config isn't changed, and an existing timestamp brought out of sync, if this is done twice

Co-authored-by: André Arnqvist <58822152+aarnq@users.noreply.github.com>
Co-authored-by: Simon Kollberg <simonkollberg@gmail.com>