bin: Allow subnet in update-ips
For OpenStack CAPI clusters allow the entire subnet instead of
individual node IPs.

This should help with autoscaling where nodes are added on demand.
Without doing this the administrator would have to run update-ips
to make sure the new nodes are allowed.

It also has the benefit of allowing non-node addresses in the subnet
such as loadbalancers.