fix: use yq4 syntax in harbor-backup
It might be the case that this script is not working correctly
right now. The chart image field says "ghcr.io/elastisys/backup-postgres:1.5.0"

If we do a `docker run --rm -it --name="random" --entrypoint=bash ghcr.io/elastisys/backup-postgres:1.5.0`

and then call `yq -V` we get: `yq (https://github.com/mikefarah/yq/) version 4.16.2`

which is already version 4, and the syntax we're using `yq r - -j` is for version 3.