bin: Do not block apply when version set to 'any' or git commit
When the 'welkin-apps-meta' configmap exists, a stricter version check was enabled which prevented running 'apply' when
running on the main branch or any untagged git commit.

This bypasses the checks if .global.ck8sVersion is set to 'any' or a random untagged git commit has been checked out IFF
`ck8s init` has been run to explicitly mark that as the current version.