bin: Add support for running parts of update-ips
It's now possible to run update-ips with the flags --enable and
--disable.

The --enable flag will run only the operation that you supply it.

The --disable flag will run all operations except the one you supply,
it can be repeated.

Not using either of the flags will run all the operations and will
work as before except that things will be running in a different
order. Functionally there should not be any difference.